# Conflicts:
#	FBOLinx.Web/FBOLinx.ServiceLayer/BusinessServices/CompanyPricingLog/CompanyPricingLogService.cs
#	FBOLinx.Web/FBOLinx.ServiceLayer/BusinessServices/FuelRequests/FuelReqService.cs
#	FBOLinx.Web/FBOLinx.ServiceLayer/BusinessServices/Groups/GroupCustomersService.cs
#	FBOLinx.Web/FBOLinx.ServiceLayer/BusinessServices/Groups/GroupFboService.cs
#	FBOLinx.Web/FBOLinx.ServiceLayer/DTO/FuelReqDto.g.cs
#	FBOLinx.Web/FBOLinx.ServiceLayer/EntityServices/CompanyPricingLogEntityService.cs
#	FBOLinx.Web/FBOLinx.Web/ClientApp/src/app/pages/customers/customers-grid/customers-grid.component.html
#	FBOLinx.Web/FBOLinx.Web/Controllers/GroupsController.cs
#	FBOLinx.Web/FBOLinx.Web/SQL/8677qtygd/CustomerAircrafts.sql

+++ FBOLinx.Web/FBOLinx.DB/Models/SWIM/SWIMFlightLeg.cs
@@ -9,7 +9,7 @@ using FBOLinx.Core.Enums;
 namespace FBOLinx.DB.Models
 {
     [Table("SWIMFlightLegs")]
-    public class SWIMFlightLeg : FBOLinxBaseEntityModel<int>
+    public class SWIMFlightLeg : FBOLinxBaseEntityModel<long>
     {
         [Column(TypeName = "varchar(50)")]
         public string Gufi { get; set; }
